Consolidate identity + trace promise propagation into HoistPromiseFactory
Merge IdentityPropagatingPromiseFactory and ContextPropagatingPromiseFactory into a single HoistPromiseFactory that captures and propagates both HoistIdentity and OTel trace context to Grails task {} workers via one delegating factory.

Installed once from HoistCoreGrailsPlugin.doWithApplicationContext() rather than from inside individual services, so neither IdentityService nor TraceContextService owns the global Promises.promiseFactory slot.

Also: noteUserAuthenticated now uses its passed-in request (via an optional final param on setIdentity), and removes an unused duplicate import from HoistFilter.
